Split LLM_main into modules
